+++ gems/aws-sdk-dsql/lib/aws-sdk-dsql.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+# WARNING ABOUT GENERATED CODE
+#
+# This file is generated. See the contributing guide for more information:
+# https://github.com/aws/aws-sdk-ruby/blob/version-3/CONTRIBUTING.md
+#
+# WARNING ABOUT GENERATED CODE
+
+
+require 'aws-sdk-core'
+require 'aws-sigv4'
+
+Aws::Plugins::GlobalConfiguration.add_identifier(:dsql)
+
+module Aws::DSQL
+  # this should get replaced by build
+
+  GEM_VERSION = '1.0.0'
+end
+
+require_relative 'aws-sdk-dsql/customizations'