+++ gems/aws-sdk-dsql/lib/aws-sdk-dsql/customizations.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'aws-sdk-dsql/customizations/auth_token_generator'